Update the U-Boot NAND boot AES decryption detection criteria
This fix only affects NAND boot.
During the loading of the U-Boot image, the loader reads the MTP USERDATA
to determine if it is a secure boot AES-encrypted image, and if so, it
proceeds to decrypt it. If it's a brand new IC, there is no issue.
However, if the IC has previously had MTP written by NuWriter, even if
it has been overwritten from secure boot to non-secure boot, the MTP
USERDATA's AES bit remains as 1. This leads to a misjudgment in this
section of the code.
This fix changes the detection criteria from only checking the MTP
USERDATA's AES bit to simultaneously checking the Secure Boot bit and
AES bit to address the aforementioned issue.

Change-Id: Ia89646821bc2c76937a2dd953e8c16f92d10c044